Refactored class name. Added documentation. Corrected indentation. Replaced shorthand variable names. Added padding between class members.
